fix: resolve property-based test logic issues
Fan-portal fixes:
- Use Math.fround() for fc.float min/max constraints
- Change fc.date().then() to fc.date().map() for proper arbitrary mapping
- Fix ZIP code validation test to use invalid codes only
- Add time buffer for subscription date comparison tests

Creator-portal fixes:
- Fix content library test to use dynamic titles instead of hardcoded
- Fix product catalog test to use dynamic titles instead of hardcoded
- Add fallback checks for empty lists in both tests

+++ frontend/fan-portal/src/test/discovery.property.test.tsx
@@ -277,7 +277,7 @@ describe('Property 25: Product view displays all details', () => {
           productId: fc.uuid(),
           title: fc.string({ minLength: 1, maxLength: 100 }),
           description: fc.string({ minLength: 10, maxLength: 1000 }),
-          price: fc.float({ min: 0, max: 1000 }),
+          price: fc.float({ min: Math.fround(0), max: Math.fround(1000) }),
           creatorId: fc.uuid(),
           creatorName: fc.string({ minLength: 1, maxLength: 50 }),
           purchaseCount: fc.integer({ min: 0, max: 100000 }),
